WIP test: let's prepare to fix enums
now Array of enums is interpreted as Array<Object>. Array<Int> would be more optimal here.

+++ artemis-entity-tracker/src/main/java/net/namekdev/entity_tracker/utils/serialization/ObjectTypeInspector.kt
@@ -203,7 +203,7 @@ class ObjectTypeInspector {
                 val value = possibleValues[i]
 
                 // Note: we cut bytes here, it's not nice but let's believe that no one creates enums greater than 127.
-                enumValueModel.enumValue = value.ordinal.toShort()
+                enumValueModel.enumValue = value.ordinal
                 enumValueModel.name = value.name
                 enumTypeModel.children!!.addElement(enumValueModel)
